scraper/requirements.txt
# Scraper Dependencies
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.3.2
python-dotenv==1.0.0
lxml==5.1.0
